- turn BaseLifecycleService.kt into a sealed class
