--- bootlaces/src/main/java/com/candroid/bootlaces/BaseLifecycleService.kt
@@ -1,75 +1,0 @@
-/*Copyright 2019 Chris Basinger
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.*/
-package com.candroid.bootlaces
-
-import android.content.Intent
-import androidx.core.app.ServiceCompat
-import androidx.lifecycle.LifecycleService
-import androidx.lifecycle.ServiceLifecycleDispatcher
-/*
-            (   (                ) (             (     (
-            )\ ))\ )    *   ) ( /( )\ )     (    )\ )  )\ )
- (   (   ( (()/(()/(  ` )  /( )\()|()/((    )\  (()/( (()/(
- )\  )\  )\ /(_))(_))  ( )(_)|(_)\ /(_))\((((_)( /(_)) /(_))
-((_)((_)((_|_))(_))   (_(_()) _((_|_))((_))\ _ )(_))_ (_))
-| __\ \ / /|_ _| |    |_   _|| || | _ \ __(_)_\(_)   \/ __|
-| _| \ V /  | || |__    | |  | __ |   / _| / _ \ | |) \__ \
-|___| \_/  |___|____|   |_|  |_||_|_|_\___/_/ \_\|___/|___/
-....................../´¯/)
-....................,/¯../
-.................../..../
-............./´¯/'...'/´¯¯`·¸
-........../'/.../..../......./¨¯\
-........('(...´...´.... ¯~/'...')
-.........\.................'...../
-..........''...\.......... _.·´
-............\..............(
-..............\.............\...
-*/
-/**
- * @author Chris Basinger
- * @email evilthreads669966@gmail.com
- * @date 11z/01/20
- * */
-abstract class BaseLifecycleService: LifecycleService() {
-    private val mDispatcher = ServiceLifecycleDispatcher(this)
-    internal var startId: Int = 0
-
-    init {
-        lifecycle.addObserver(BootServiceState)
-    }
-
-    override fun getLifecycle() = mDispatcher.lifecycle
-
-    override fun onCreate() {
-        mDispatcher.onServicePreSuperOnCreate()
-        super.onCreate()
-    }
-
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        mDispatcher.onServicePreSuperOnStart()
-        super.onStartCommand(intent, flags, startId)
-        this.startId = startId
-        return START_STICKY
-    }
-
-    override fun onDestroy() {
-        mDispatcher.onServicePreSuperOnDestroy()
-        if(BootServiceState.isForeground())
-            ServiceCompat.stopForeground(this, ServiceCompat.STOP_FOREGROUND_REMOVE)
-        lifecycle.removeObserver(BootServiceState)
-        stopSelfResult(startId)
-        super.onDestroy()
-    }
-}
